Refactored code into another component
