+++ to-do-list/src/ListItem.js
@@ -1,0 +1,5 @@
+export const ListItem = ({items}) => {
+    return (
+        <ul>{items.map((item, index) => <li key={index}>{item}</li>)}</ul>
+    )
+}
